[bugfix] Changed query to use criteria instead of source
Doctrine-dbal implementation of PHPCR does not allow to select FROM a
mixin type. This fix moves the node type specification to the criteria.

Note that this should be reverted when doctrine-dbal does support it, as
it can be more performant (but current benchmarks do not indicate a
performance penalty on jackrabbit).