Use non-deprecated locale method
I managed to track line `191` down as a source of Chrome `ERR_EMPTY_RESPONSE` errors. Changing it to the newer method stops the error in my case. I'm not sure what in my data is causing it currently but it also happens when you have no pages except the home page and several locales in the same language (GB, US, AU) for example.

First tracked it to the `getContentPath` method in `ContentPathTwigExtension.php` and followed the trail into `WebspaceManager.php`. My xDebug isn't working currently but this fixes whatever conditions led to the error.

This sort of locale switch causes the start of the problem:
```
{% set size = size | default('') %}
<ul class="locale-switch">
    {% for locale, url in urls %}
        <li{% if request.locale == locale %} class="active"{% endif %}>
            <a href="{{ sulu_content_path((url?:'/'), request.webspaceKey, locale) }}">
                {% include 'components/atoms/flag/' ~ locale ~ '.html.twig' with { "size": size } %}
            </a>
        </li>
    {% endfor %}
</ul>
```