Correct extension stripping (#3757)
* Correct extension stripping

The removal of the extension of formats should only take place when
the format is present as extension in the url. (#3755)

* Test and changelog for correct extension stripping

Add tests and a changelog entry for the extension stripping.
The removal of the extension of formats should only take place when
the format is present as extension in the url. (#3755)