Change: always show content type in collapsed block
Previously, the content type of a collapsed block would only be shown if
there was more than 1 type defined for the given block. However, for
editors that makes it impossible to see what kind of content is in that
spot. This change will make it so the content type is always shown.