Remove duplicate event handler
While testing, I noticed that an onKeyDown handler was defined on these buttons in addition to the onClick handler. This is unnecessary because native button elements are already triggered by the click event. This redundancy is interesting because a wrong event key string was used for the space button, so these buttons already fall back to the onClick event handler when the space key is pressed.