Remove duplicate event handler (#6456)
* Remove duplicate event handler

While testing, I noticed that an onKeyDown handler was defined on these buttons in addition to the onClick handler. This is unnecessary because native button elements are already triggered by the click event. This redundancy is interesting because a wrong event key string was used for the space button, so these buttons already fall back to the onClick event handler when the space key is pressed.

* Remove duplicate tests

These four tests duplicate the native click event and are therefore no longer necessary.

* Add missing type attribute for the SelectOption

Co-authored-by: Alexander Schranz <alexander@sulu.io>

* Add missing type attribute for the SelectOption

Co-authored-by: Alexander Schranz <alexander@sulu.io>

* Update snapshot tests

* Readding listener for enter only

* Readding on enter listener

Co-authored-by: Alexander Schranz <alexander@sulu.io>