Adding admin resource URLs to FOSJsRouter automatically
A new compiler pass collects all configured routes for admin resources and adds them to the argument of the `ExposedRoutesExtractor` service of the FOSJsRouterBundle. This way it is not required to follow a naming convention for those routes or add them to the bundle's config manually anymore.

Manual additions of resource routes to the FOSJsBundle are obsolete, now and have been removed.

For details see the RFC on https://github.com/sulu/sulu/issues/6886