Adding types to Symfony Config
In Symfony 6 not having this type is deprecated and it can't hurt in 5.4 either. This is not a BC break as Return Type narrowing (the implementation of an interface can have a narrower return type than the interface implementing it) was introduced in 7.4 and not returning a TreeBuilder crashes the application anyways.