Limit symlink-dir version to <6.0
The symlink-dir package is required in the `symlink-vendor-directory.js`. Version 6.0 was just released and it requires node >= 18.12. But node 18 comes with npm 9, which is currently not supported by Sulu.

By the limititation of symlink-dir's version to something lower than 6.0 we make sure that npx does not load an incompatible version.