Deprecate defining different host and scheme in kernel parameters "router.request_context.*" and "framework.router.default_uri"
The webspace manager, which is configured in the core bundle, relies on the request context container parameters `router.request_context.host` and `router.request_context.scheme` as fallback when no actual request object is available. This might happen during cache warming in a functional test.

But if the request context is configured by Symfony's `framework.router.default_uri` setting like it is described in their [docs](https://symfony.com/doc/current/routing.html#generating-urls-in-commands), the mentioned container parameters still uses the defaults (localhost and http). There is a issue which points this out https://github.com/symfony/symfony/issues/53919.

This leads to wrong URL generation in a functional test (kernel test case) when the cache is warmed.

Therfore a deprecation message will be triggered, now, telling the user that the Symfony config `framework.router.default_uri` needs to have the same host and scheme like the mentioned parameters.

Starting from Sulu 3.0 only the URI defined in `framework.router.default_uri` will be respected.

This change fixes this issue (described in #7280).